+++ wp-content/themes/intranet/assets/source/js/User/FacebookProfileSync.js
@@ -1,0 +1,46 @@
+Intranet = Intranet || {};
+Intranet.User = Intranet.User || {};
+
+Intranet.User.FacebookProfileSync = (function ($) {
+    function FacebookProfileSync() {
+
+    }
+
+    FacebookProfileSync.prototype.getDetails = function() {
+        $('.fb-login-container .fb-login-button').hide();
+        $('.fb-login-container').append('<div class="loading loading-red"><div></div><div></div><div></div><div></div></div>');
+
+        FB.api('/me', {fields: 'birthday, location'}, function (details) {
+            this.saveDetails(details);
+        }.bind(this));
+    };
+
+    FacebookProfileSync.prototype.saveDetails = function(details) {
+        var data = {
+            action: 'sync_facebook_profile',
+            details: details
+        };
+
+        $.post(ajaxurl, data, function (response) {
+            if (response !== '1') {
+                $('.fb-login-container .loading').remove();
+                $('.fb-login-container').append('<div class="notice warning">Facebook details did not sync due to an error</div>');
+
+                return false;
+            }
+
+            $('.fb-login-container .loading').remove();
+            $('.fb-login-container').append('<div class="notice success">Facebook details synced to your profile</div>');
+
+            return true;
+        });
+    };
+
+    return new FacebookProfileSync();
+
+})(jQuery);
+
+
+function facebookProfileSync() {
+    Intranet.User.FacebookProfileSync.getDetails();
+}