+++ wp-content/themes/intranet/library/Public/User.php
@@ -108,3 +108,66 @@ if (!function_exists('municipio_intranet_get_administration_unit_name')) {
         return \Intranet\User\AdministrationUnits::getAdministrationUnit($id);
     }
 }
+
+if (!function_exists('municipio_intranet_user_has_birthday')) {
+    /**
+     * Check if user has birthday today
+     * @param  mixed $user User id or login name, default is current logged in user
+     * @return string
+     */
+    function municipio_intranet_user_has_birthday($user = null)
+    {
+        if (is_null($user)) {
+            $user = wp_get_current_user();
+        } elseif (is_numeric($user)) {
+            $user = get_user_by('ID', $user);
+        } elseif (is_string($user)) {
+            $user = get_user_by('slug', $user);
+        }
+
+        if (get_the_author_meta('user_hide_birthday') == 1) {
+            return false;
+        }
+
+        $birthday = false;
+        if (!get_the_author_meta('user_birthday') || !get_the_author_meta('user_birthday')['year'] || !get_the_author_meta('user_birthday')['month'] || !get_the_author_meta('user_birthday')['day']) {
+            return false;
+        }
+
+        if (get_the_author_meta('user_birthday')['month'] == date('m') && get_the_author_meta('user_birthday')['day'] == date('d')) {
+            return true;
+        }
+
+        return false;
+    }
+}
+
+if (!function_exists('municipio_intranet_user_birthday')) {
+    /**
+     * Check if user has birthday today
+     * @param  mixed $user User id or login name, default is current logged in user
+     * @return string
+     */
+    function municipio_intranet_user_birthday($user = null)
+    {
+        if (is_null($user)) {
+            $user = wp_get_current_user();
+        } elseif (is_numeric($user)) {
+            $user = get_user_by('ID', $user);
+        } elseif (is_string($user)) {
+            $user = get_user_by('slug', $user);
+        }
+
+        if (get_the_author_meta('user_hide_birthday') == 1) {
+            return false;
+        }
+
+        if (!get_the_author_meta('user_birthday') || !get_the_author_meta('user_birthday')['year'] || !get_the_author_meta('user_birthday')['month'] || !get_the_author_meta('user_birthday')['day']) {
+            return false;
+        }
+
+        $bday = get_the_author_meta('user_birthday')['year'] . '-' . get_the_author_meta('user_birthday')['month'] . '-' . get_the_author_meta('user_birthday')['day'];
+
+        return mysql2date('d F Y', $bday);
+    }
+}